Check if sleepimage exists before deleting it.

+++ osx/os.sh
@@ -17,7 +17,7 @@ defaults write NSGlobalDomain NSDocumentSaveNewDocumentsToCloud -bool false
 # Disable hibernation (speeds up entering sleep mode)
 sudo pmset -a hibernatemode 0
 # Remove the sleep image file to save disk space
-sudo rm /private/var/vm/sleepimage
+[[ -f /private/var/vm/sleepimage ]] && sudo rm /private/var/vm/sleepimage
 # Create a zero-byte file instead…
 sudo touch /private/var/vm/sleepimage
 # …and make sure it can’t be rewritten
